requirements.txt
fastapi
uvicorn
sqlalchemy
openai
python-multipart
python-dotenv
pypdf
pdfplumber
pdf2image
pytesseract
Pillow
python-dotenv